Weather_App/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pillow==12.0.0